# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Attendance Gateway API",
    "version": "1.4.0",
    "description": "Zero-Trust Cryptographic Attendance Gateway.\n\n**Authentication**\n- Most client endpoints accept a shared client API key: header `X-Api-Key: ag_...` (mint from the Admin console → API Keys).\n- Role-protected endpoints (professor/admin) use a JWT: `Authorization: Bearer <token>` from `/api/v1/admin/login` or `/api/v1/auth/login`.\n- Global security: `ApiKey OR bearer`; public operations (health, admin login) override with `[]`.\n\n**Building a client?** See the **Client Integration Guide** (QR / 3-second metronome protocol, biometric Gate 2 ordering, timestamp sealing, 250ms window): https://git.atmyhome.tech/het/attendance-gateway/src/branch/main/docs/CLIENT_INTEGRATION_GUIDE.md\n\n**Error payloads:** every 4xx/5xx returns the envelope `{success:false,error:{code,message,latency_ms?,details?}}`. Full code table + client UI actions: **Error Dictionary** — https://git.atmyhome.tech/het/attendance-gateway/src/branch/main/docs/ERROR_DICTIONARY.md (schema: `ErrorEnvelope`).\n\n**To get an API key:** sign in to the Admin console (https://admin.atmyhome.tech) → **API Keys** → Generate Key."
  },
  "servers": [
    {
      "url": "https://api.atmyhome.tech"
    }
  ],
...
  "security": [
    {
      "ApiKey": []
    },
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Api-Key",
        "description": "Shared client API key (mint from the Admin console)."
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/api/v1/admin/students": {
      "get": {
        "tags": [
          "Admin - Core"
        ],
        "summary": "List students",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Admin - Core"
        ],
        "summary": "Create a student",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "roll_no": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "name": {
                    "type": "string"
                  },
                  "division_id": {
                    "type": "string"
                  },
                  "batch_id": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
...
    "/api/v1/admin/courses": {
      "get": {
        "tags": [
          "Admin - Core"
        ],
        "summary": "List courses",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Admin - Core"
        ],
        "summary": "Create a course",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "course_code": {
                    "type": "string"
                  },
                  "title": {
                    "type": "string"
                  },
                  "division_id": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
...
    "/api/v1/admin/students/{uuid}": {
      "get": {
        "tags": [
          "Admin - Core"
        ],
        "summary": "Get student detail",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Admin - Core"
        ],
        "summary": "Update a student",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "roll_no": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "name": {
                    "type": "string"
                  },
                  "division_id": {
                    "type": "string"
                  },
                  "batch_id": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Admin - Core"
        ],
        "summary": "Delete a student",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v1/admin/students/{uuid}/reset-device": {
      "post": {
        "tags": [
          "Admin - Core"
        ],
        "summary": "Reset device bindings (Unbind hardware)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v1/admin/students/{uuid}/rotate-hmac": {
      "post": {
        "tags": [
          "Admin - Core"
        ],
        "summary": "Rotate student HMAC signature key",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v1/admin/students/{uuid}/forgot-password": {
      "post": {
        "tags": [
          "Admin - Core"
        ],
        "summary": "Force clear student password (admin initiated reset)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v1/admin/courses/{code}": {
      "put": {
        "tags": [
          "Admin - Core"
        ],
        "summary": "Update a course",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "course_code": {
                    "type": "string"
                  },
                  "title": {
                    "type": "string"
                  },
                  "division_id": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Admin - Core"
        ],
        "summary": "Delete a course",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  }
}